backend/requirements.txt
fastapi
uvicorn[standard]
ffmpeg-python
yolov5
torch
torchvision
openai